aes: Move `Counter` & `Iv` to their own submodule.
Move all the counter management to a submodule that can, in the future,
be encapsulated from the rest of the AES and AES-GCM logic.

`git difftool HEAD^1:src/aead/aes.rs src/aead/aes/counter.rs`
